% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromasc.R
\name{ChromaSC}
\alias{ChromaSC}
\title{ChromaSC}
\usage{
ChromaSC(seurat_obj, traj_vec, window_size, step_size, gmt_path)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{traj_vec}{A vector of continuous values named by a subset of seurat_obj@cell.names}

\item{window_size}{A double in (0.0, 1.0) giving the fraction of the trajectory to use for each window}

\item{step_size}{A double in (0.0, 1.0) giving the fraction the window should be shifted each iteration}

\item{gmt_path}{A path to a GMT file with one or more gene sets}
}
\value{
A gene-sets-by-windows data.frame with the enrichment values for each gene set across all windows
}
\description{
ChromaSC
}
