% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileutils.R
\name{load_gene_sets}
\alias{load_gene_sets}
\title{load_gene_sets}
\usage{
load_gene_sets(path)
}
\arguments{
\item{path}{A path to a .gmt}
}
\description{
This function takes a path to a GMT format file and returns a named list of
character vectors corresponding to each gene set.
For more on the GMT format, see the
\href{http://software.broadinstitute.org/cancer/software/genepattern/file-formats-guide#GMT}{GenePattern documentation}
}
