% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromasc.R
\name{run_enrichment}
\alias{run_enrichment}
\title{run_enrichment}
\usage{
run_enrichment(expr_mat, gene_sets, null_params)
}
\arguments{
\item{expr_mat}{An expression matrix}

\item{gene_sets}{A gene set object loaded by `load_gene_sets()`}
}
\description{
Hi
}
