% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_distr_params.R
\name{compute_null_params}
\alias{compute_null_params}
\title{compute_null_params}
\usage{
compute_null_params(expr_mat, gene_sets, n_random_cells = 50)
}
\arguments{
\item{expr_mat}{A sparse Matrix object}

\item{gene_sets}{Gene sets loaded with `load_gene_sets()`}

\item{n_random_cells}{The number of random cells to generate to calculate the null parameters
for each combination of cell size and gene set}
}
\description{
This function computes the null distribution parameters for the input dataset. Use
`save_null_params()` and `load_null_params()` to skip this step for future analyses with the same
gene sets and data.
}
