% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_distr_params.R
\name{compute_null_params}
\alias{compute_null_params}
\title{compute_null_params}
\usage{
compute_null_params(seurat_obj, gene_sets)
}
\arguments{
\item{seurat_obj}{A seurat object}

\item{gene_sets}{Gene sets loaded with `load_gene_sets()`}
}
\description{
This function computes the null distribution parameters for the input dataset. Use
`save_null_params()` and `load_null_params()` to skip this step for future analyses with the same
gene sets and data.
}
