% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromasc.R
\name{run_enrichment}
\alias{run_enrichment}
\title{run_enrichment}
\usage{
run_enrichment(seurat_obj, traj_vecs, gmt_path, window_size, step_size)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{traj_vecs}{A list of vectors of continuous values named by a subset of seurat_obj@cell.names}

\item{gmt_path}{A path to a GMT file with one or more gene sets}

\item{window_size}{The fraction of pseudotime to use for the sliding window}

\item{step_size}{The fraction of pseudotime values to move the sliding window on each iteration}
}
\description{
Hi
}
